modified new API endpoints to separate data access layer & match rest of repo
